heartbeat.mp3
This file is the sample sound file that I am working with to create an accurate heartbeat monitor. You will need this file for the matlab code to run properly (and put it within your matlab folder or path). The total file size is 2.29 Mb, but I am only playing through and displaying the first 100000 data points in Matlab. (There are only 4 sample heart beats that I am currently working with).